Move search logic into searches.go
